Reuse Location for other metrics
