+++ eslint-bridge/tests/server.test.ts
@@ -20,31 +20,39 @@ const expectedResponse = {
   highlightedSymbols: [],
   highlights: [
     {
-      startLine: 1,
-      startCol: 0,
-      endLine: 1,
-      endCol: 2,
+      location: {
+        startLine: 1,
+        startCol: 0,
+        endLine: 1,
+        endCol: 2,
+      },
       textType: "KEYWORD",
     },
     {
-      startLine: 1,
-      startCol: 10,
-      endLine: 1,
-      endCol: 12,
+      location: {
+        startLine: 1,
+        startCol: 10,
+        endLine: 1,
+        endCol: 12,
+      },
       textType: "CONSTANT",
     },
     {
-      startLine: 1,
-      startCol: 14,
-      endLine: 1,
-      endCol: 18,
+      location: {
+        startLine: 1,
+        startCol: 14,
+        endLine: 1,
+        endCol: 18,
+      },
       textType: "KEYWORD",
     },
     {
-      startLine: 1,
-      startCol: 19,
-      endLine: 1,
-      endCol: 21,
+      location: {
+        startLine: 1,
+        startCol: 19,
+        endLine: 1,
+        endCol: 21,
+      },
       textType: "CONSTANT",
     },
   ],
@@ -59,66 +67,84 @@ const expectedResponse = {
   },
   cpdTokens: [
     {
-      startLine: 1,
-      startCol: 0,
-      endLine: 1,
-      endCol: 2,
+      location: {
+        startLine: 1,
+        startCol: 0,
+        endLine: 1,
+        endCol: 2,
+      },
       image: "if",
     },
     {
-      startLine: 1,
-      startCol: 3,
-      endLine: 1,
-      endCol: 4,
+      location: {
+        startLine: 1,
+        startCol: 3,
+        endLine: 1,
+        endCol: 4,
+      },
       image: "(",
     },
     {
-      startLine: 1,
-      startCol: 4,
-      endLine: 1,
-      endCol: 8,
+      location: {
+        startLine: 1,
+        startCol: 4,
+        endLine: 1,
+        endCol: 8,
+      },
       image: "true",
     },
     {
-      startLine: 1,
-      startCol: 8,
-      endLine: 1,
-      endCol: 9,
+      location: {
+        startLine: 1,
+        startCol: 8,
+        endLine: 1,
+        endCol: 9,
+      },
       image: ")",
     },
     {
-      startLine: 1,
-      startCol: 10,
-      endLine: 1,
-      endCol: 12,
+      location: {
+        startLine: 1,
+        startCol: 10,
+        endLine: 1,
+        endCol: 12,
+      },
       image: "42",
     },
     {
-      startLine: 1,
-      startCol: 12,
-      endLine: 1,
-      endCol: 13,
+      location: {
+        startLine: 1,
+        startCol: 12,
+        endLine: 1,
+        endCol: 13,
+      },
       image: ";",
     },
     {
-      startLine: 1,
-      startCol: 14,
-      endLine: 1,
-      endCol: 18,
+      location: {
+        startLine: 1,
+        startCol: 14,
+        endLine: 1,
+        endCol: 18,
+      },
       image: "else",
     },
     {
-      startLine: 1,
-      startCol: 19,
-      endLine: 1,
-      endCol: 21,
+      location: {
+        startLine: 1,
+        startCol: 19,
+        endLine: 1,
+        endCol: 21,
+      },
       image: "42",
     },
     {
-      startLine: 1,
-      startCol: 21,
-      endLine: 1,
-      endCol: 22,
+      location: {
+        startLine: 1,
+        startCol: 21,
+        endLine: 1,
+        endCol: 22,
+      },
       image: ";",
     },
   ],
